AST+Parser: Decrease size of NodeId to u32 to save 4-bytes per node

+++ crates/pxp-ast/src/lib.rs
@@ -63,7 +63,7 @@ pub mod variables;
 pub mod data_type;
 
 pub type Block = Vec<Statement>;
-pub type NodeId = usize;
+pub type NodeId = u32;
 
 #[derive(Debug, PartialEq, Eq, Clone)]
